Updated README.md Fixed bug allowing 'views\admins\plans' `trial` field to accept non-numeric values. Fix inconsistency of database `id` incremental keys.  All `id` are now 64-bit size. Added 'Site' model and database migration. Added 'plan_id' to 'subscription' table.

+++ database/migrations/2018_01_29_140758_create_confirmation_tokens_table.php
@@ -4,22 +4,27 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateConfirmationTokensTable extends Migration
-{
+class CreateConfirmationTokensTable extends Migration {
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
+    public function up() {
         Schema::create('confirmation_tokens', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigIncrements('id');
+            $table
+                ->integer('user_id')
+                ->unsigned()
+                ->index();
             $table->string('token');
             $table->timestamp('expires_at');
 
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            $table
+                ->foreign('user_id')
+                ->references('id')
+                ->on('users')
+                ->onDelete('cascade');
         });
     }
 
@@ -28,8 +33,7 @@ class CreateConfirmationTokensTable extends Migration
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         Schema::dropIfExists('confirmation_tokens');
     }
 }
