+++ database/migrations/2018_01_29_154234_create_role_permissions_table.php
@@ -4,22 +4,34 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateRolePermissionsTable extends Migration
-{
+class CreateRolePermissionsTable extends Migration {
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
+    public function up() {
         Schema::create('role_permissions', function (Blueprint $table) {
-            $table->integer('role_id')->unsigned()->index();
-            $table->integer('permission_id')->unsigned()->index();
+            $table
+                ->bigInteger('role_id')
+                ->unsigned()
+                ->index();
+            $table
+                ->bigInteger('permission_id')
+                ->unsigned()
+                ->index();
             $table->timestamps();
 
-            $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-            $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
+            $table
+                ->foreign('role_id')
+                ->references('id')
+                ->on('roles')
+                ->onDelete('cascade');
+            $table
+                ->foreign('permission_id')
+                ->references('id')
+                ->on('permissions')
+                ->onDelete('cascade');
 
             $table->primary(['role_id', 'permission_id']);
         });
@@ -30,8 +42,7 @@ class CreateRolePermissionsTable extends Migration
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         Schema::dropIfExists('role_permissions');
     }
 }
