+++ database/migrations/2018_04_27_160948_create_projects_table.php
@@ -4,23 +4,31 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateProjectsTable extends Migration
-{
+class CreateProjectsTable extends Migration {
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
+    public function up() {
         Schema::create('projects', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('company_id')->unsigned()->index();
+            $table->bigIncrements('id');
+            $table
+                ->integer('company_id')
+                ->unsigned()
+                ->index();
             $table->string('name', 250);
-            $table->string('slug', 250)->unique()->index();
+            $table
+                ->string('slug', 250)
+                ->unique()
+                ->index();
             $table->timestamps();
 
-            $table->foreign('company_id')->references('id')->on('companies')->onDelete('cascade');
+            $table
+                ->foreign('company_id')
+                ->references('id')
+                ->on('companies')
+                ->onDelete('cascade');
         });
     }
 
@@ -29,8 +37,7 @@ class CreateProjectsTable extends Migration
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         Schema::dropIfExists('projects');
     }
 }
